refactor: moved time chart logic to a seperate utils file
