+++ src/pages/BudgetBreakdown/components/TimeChart/TimeChart.tsx
@@ -6,15 +6,12 @@ import localizedFormat from 'dayjs/plugin/localizedFormat';
 import { Box } from '@mui/material';
 
 import { useAppSelector } from '../../../../hooks/ReduxHookWrappers';
+// import useContentWidth from '../../../../hooks/useContentWidth';
 
 import { getCategoryOrderedDataById } from '../../../../redux/selectors/categorySelectors';
 
-import type {
-    IProps,
-    ISortedByCategory,
-    ISortedByCategoryRow,
-} from './TimeChart.types';
-import useContentWidth from '../../../../hooks/useContentWidth';
+import type { IProps } from './TimeChart.types';
+import { generateTimeChartSeries } from './TimeChartUtils';
 
 dayjs.extend(localizedFormat);
 
@@ -31,76 +28,18 @@ const TimeChart: FC<IProps> = ({
 }) => {
     const categories = useAppSelector(getCategoryOrderedDataById);
 
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { contentWidth } = useContentWidth();
+    // const { contentWidth } = useContentWidth();
 
-    const series = useMemo(() => {
-        const dates: number[] = [];
-        const endDateJs = endDate;
-        let date = startDate;
-
-        while (date.valueOf() <= endDateJs.valueOf()) {
-            dates.push(date.valueOf());
-            date = date.add(1, 'day');
-        }
-
-        const sortedByCategory = filteredTransactions.reduce(
-            (acc: ISortedByCategory, each) => {
-                if (each.category_id) {
-                    if (!(each.category_id in acc)) {
-                        const foundCategory = categories[each.category_id];
-                        acc[each.category_id] = {
-                            label:
-                                foundCategory?.label ||
-                                `Category ID ${each.category_id}`,
-                            id: each.category_id,
-                            transactions: {},
-                        };
-                    }
-                    const dateInt = dayjs(each.date).valueOf();
-                    if (!(dateInt in acc[each.category_id].transactions)) {
-                        acc[each.category_id].transactions[dateInt] = [];
-                    }
-                    acc[each.category_id].transactions[dateInt].push(each);
-                }
-                return acc;
-            },
-            {},
-        );
-
-        const nextSeries = Object.values(sortedByCategory).map(
-            (seriesItem: ISortedByCategoryRow) => {
-                const { data } = dates.reduce(
-                    (
-                        accumulator: {
-                            data: { x: string; y: number }[];
-                            total: number;
-                        },
-                        nextDate,
-                    ) => {
-                        if (nextDate in seriesItem.transactions) {
-                            accumulator.total += seriesItem.transactions[
-                                nextDate
-                            ].reduce((a, e) => a + e.debit - e.credit, 0);
-                        }
-                        accumulator.data.push({
-                            x: dayjs(nextDate).toISOString(),
-                            y: accumulator.total,
-                        });
-                        return accumulator;
-                    },
-                    { data: [], total: 0 },
-                );
-
-                return {
-                    name: seriesItem.label,
-                    data,
-                };
-            },
-        );
-
-        return nextSeries;
-    }, [categories, endDate, filteredTransactions, startDate]);
+    const series = useMemo(
+        () =>
+            generateTimeChartSeries(
+                categories,
+                endDate,
+                filteredTransactions,
+                startDate,
+            ),
+        [categories, endDate, filteredTransactions, startDate],
+    );
 
     return (
         <Box
